fix: Multiple type and compiler compatibility fixes
Zombie.cpp:
- Fix min(mPhaseCounter, 40.0f) with float cast
- Fix enum comparison with explicit int cast

DataSync.cpp:
- Add #include <algorithm>
- Use std::min instead of min

Music.cpp:
- Fix delete[] on void* with static_cast<char*>

PakInterface.h:
- Fix format specifier %d to %lu for DWORD

+++ Lawn/Zombie.cpp
@@ -5497,7 +5497,7 @@ void Zombie::DrawBungeeTarget(Graphics* g)
     if (mZombiePhase == ZombiePhase::PHASE_BUNGEE_HIT_OUCHY || mZombiePhase == ZombiePhase::PHASE_BUNGEE_RISING)
         return;
     
-    if (mRelatedZombieID != ReanimationID::REANIMATIONID_NULL)
+    if ((int)mRelatedZombieID != (int)ReanimationID::REANIMATIONID_NULL)
         return;
     
     ZombieDrawPosition aDrawPos;
@@ -5915,7 +5915,7 @@ void Zombie::GetDrawPos(ZombieDrawPosition& theDrawPos)
         }
         else
         {
-            float aHeightLimit = min(mPhaseCounter, 40.0f);
+            float aHeightLimit = min((float)mPhaseCounter, 40.0f);
             theDrawPos.mClipHeight = theDrawPos.mBodyY + aHeightLimit;
         }
